web/requirements.txt
Django
djangorestframework
gunicorn
zmq
pytz
